requirements.txt
boto3
pycurl
lmdb
dnspython
stem
pyyaml
pillow
ray
boto3
paramiko
opencv-python
reppy
tensorboardX